requirements.txt
babel==2.9.0
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==0.14.3
flask_sqlalchemy==2.4.4
flask-migrate==3.1.0
python-dotenv==0.19.1
psycopg2==2.9.1
